Global Operations base support
Basic Global Operations support has been added. This game needs a ton of additional checks, the one that stands out is the multiple StaticSunLight objects.

+++ Assets/Scripts/DATv70/WorldObjects.cs
@@ -45,7 +45,7 @@ public class LithObject
 
 public class VertexColor
 {
-    public Int16 nVerts;
+    public UInt32 nVerts;
     public byte red;
     public byte green;
     public byte blue;
@@ -53,7 +53,7 @@ public class VertexColor
 
 public class DiskRelVert
 {
-    public Int16 nRelVerts;
+    public Int32 nRelVerts;
 }
 
 public class Leafs
